third_party/WebKit/LayoutTests/platform/linux/virtual/outofblink-cors-http/http/tests/security/img-redirect-to-crossorigin-credentials-expected.txt

CONSOLE ERROR: Redirect from 'http://localhost:8000/security/resources/abe-allow-star.php?count=2&Cookie=cross' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Redirect from 'http://localhost:8000/security/resources/abe-allow-star.php?count=3' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Redirect from 'http://localhost:8000/security/resources/abe-allow-credentials.php?count=4&Cookie=cross' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Same origin destination. crossOrigin set to anonymous 
PASS Same origin destination. crossOrigin set to use-credentials 
FAIL Cross origin destination. crossOrigin set to anonymous. Response includes wildcard Access-Control-Allow-Origin. promise_test: Unhandled rejection with value: "Image not loaded unexpectedly"
PASS Cross origin destination. crossOrigin set to use-credentials. Response includes wildcard Access-Control-Allow-Origin. Fails due to absence of Access-Control-Allow-Credentials. 
FAIL Cross origin destination. crossOrigin set to use-credentials. Response includes Access-Control-Allow-Credentials. promise_test: Unhandled rejection with value: "Image not loaded unexpectedly"
Harness: the test ran to completion.
